DUI编辑框初步，增加分割布局，图片框初步，字符串特性增加指定长度比较，选择夹适配暗色

+++ eck/CFile.h
@@ -116,7 +116,7 @@ public:
 			return *this << U::CharTerminatingNull();
 		else
 		{
-			Write(Buf.Data(), Cch2Cb(Buf.Size()));
+			Write(Buf.Data(), (DWORD)Cch2Cb(Buf.Size()));
 			return *this;
 		}
 	}
